+++ data/ascension/ascensionLogic.js
@@ -157,6 +157,7 @@ Game.ascension = (function () {
         if (Game.research) Game.research.initialise();
         if (Game.structures) Game.structures.initialise();
         if (Game.worldCycle && Game.worldCycle.resetForAscension) Game.worldCycle.resetForAscension();
+        if (Game.bosses) { Game.bosses.defeated = {}; Game.bosses.resetForAscension(); }
 
         // Reset the three primary MycoVerse currencies while leaving the legacy Space Company engine intact.
         if (typeof window.wood !== "undefined") window.wood = 0;